updated migrations

+++ db/migrate/005_remove_hideable_cf.rb
@@ -1,7 +1,7 @@
 class RemoveHideableCf < ActiveRecord::Migration
   def change
 
-    drop_table :akquinet_time_entry_custom_fields_activities
+    drop_table :akquinet_time_entry_custom_fields_activities if table_exists?(:akquinet_time_entry_custom_fields_activities)
 
   end
 end
